Major Universal Changes
--Swaps cin & cout with scanf & printf respectively. In execution, decreases memory overhead while providing improved performance. Matches the change from fstream to FILE*.

--Changes the global functions that previously returned the result of menu choice as integers to instead return characters representing the results. This should make it easier to decipher what choice does what when utilized in code.

--Additionally, changes vectorValueInsert to listValueInsert and made it non-template with only size_t.

+++ CHC/CHC_Objects/Chart.h
@@ -116,7 +116,7 @@ public:
 	size_t add(Note*);
 	void add_back(Note*);
 	bool resize(long numElements, char type = 't');
-	bool remove(size_t, char type = 't', unsigned long extra = 0);
+	bool remove(size_t, char type = 't', size_t extra = 0);
 	void clearTracelines();
 	void clearPhrases();
 	void clearGuards();
